requirements.txt
streamlit
pandas
langchain
langchain_community
openai
python-dotenv
plotly 
seaborn
FPDF
BytesIO